requirements.txt
pandas>=2.0.0
textblob>=0.17.0
pytest>=7.4.0
pytest-cov>=4.1.0
ruff>=0.1.0
python-dotenv>=1.0.0
mypy>=1.5.0
